+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## 0.19.0
 * Add new high level API
-* Deprecated: use `hasPairing` instead of `isPaired`
-* Deprecated: use `write` instead of `writeValue`
 * Deprecated: use `read` instead of `readValue`
+* Deprecated: use `write` instead of `writeValue`
+* Deprecated: Use `subscribeNotifications` or `subscribeIndications` or `unSubscribe` instead of `setNotifiable`
+* Deprecated: use `hasPairing` instead of `isPaired`
 
 ## 0.19.0
 * Get and prefer advertised name of scanned devices on Apple